Make an issue from values in household-level microdata

Make an issue from values in household-level microdata

make_issue(df, where, type = 1, desc, comment, issue_vars)

Arguments

df
Data frame of microdata

where
Logical expression that if TRUE indicates an issue.

type
Type of issue. Values are as follows: c(Reject = 1, Comment = 2, Review = 4)

desc
Character. Short description for survey managers to understand the issue.

comment
Character. Longer message for interviewers to understand the issue and what to do about it.

issue_vars
Character. Regular expression to describe variables involved in issue.

Value

Data frame of issues with the following columns:

interview__id

interview__key

issue_type

issue_desc

issue_comment

issue_vars

issue_loc